requirements.txt
google-generativeai
google-search-results==2.4.0
python-dotenv
streamlit
requests
rich
markdown
kaggle